move batch implementation to own file; move some related funcs to src/tensor.hh; add AVX2 SIMD functions
